+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.2.0 (2019-XX-XX)
 
+- Added possibility to use `aiodogstatsd.Client` as a context manager.
+
 ## 0.1.0 (2019-02-10)
 
 - Initial release.